% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oskeyring-package.R
\docType{package}
\name{oskeyring-package}
\alias{oskeyring}
\alias{oskeyring-package}
\title{oskeyring: Raw System Credential Store Access from R}
\description{
Aims to support all features of the system credential store, including non-portable ones. Supports 'Keychain' on 'macOS', and 'Credential Manager' on 'Windows'. See the 'keyring' package if you need a portable 'API'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/oskeyring#readme}
  \item \url{https://r-lib.github.io/oskeyring/}
  \item Report bugs at \url{https://github.com/r-lib/oskeyring/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
