% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macos.R
\name{macos_keychain}
\alias{macos_keychain}
\alias{macos_item_classes}
\alias{macos_item}
\alias{macos_item_add}
\alias{macos_item_search}
\alias{macos_item_update}
\alias{macos_item_delete}
\alias{macos_keychain_create}
\alias{macos_keychain_list}
\alias{macos_keychain_delete}
\alias{macos_keychain_lock}
\alias{macos_keychain_unlock}
\alias{macos_keychain_is_locked}
\alias{macos_keychain_file}
\title{Query and manipulate the macOS Keychain}
\usage{
macos_item_classes()

macos_item(value, attributes = list(), class = "generic_password")

macos_item_add(item, keychain = NULL)

macos_item_search(
  class = "generic_password",
  attributes = list(),
  match = list(),
  keychain = NULL
)

macos_item_update(
  class = "generic_password",
  attributes = list(),
  match = list(),
  update = list(),
  keychain = NULL
)

macos_item_delete(
  class = "generic_password",
  attributes = list(),
  match = list(),
  keychain = NULL
)

macos_keychain_create(keychain, password = NULL)

macos_keychain_list()

macos_keychain_delete(keychain)

macos_keychain_lock(keychain = NULL)

macos_keychain_unlock(keychain = NULL, password = NULL)

macos_keychain_is_locked(keychain = NULL)

macos_keychain_file(keychain = NULL)
}
\arguments{
\item{value}{Value of the item, a password, key or certificate. It must
a raw vector or a string. If it is a string, then it is converted to
UTF-8.}

\item{attributes}{Narrow the search by indicating the attributes that
the found item or items should have.}

\item{class}{Type of items to search, see \code{\link[=macos_item_classes]{macos_item_classes()}} for
possible values.}

\item{item}{Keychain item, creted via \code{\link[=macos_item]{macos_item()}} or returned
by oskeyking itself.}

\item{keychain}{Keychain to use. \code{NULL} means the default one.}

\item{match}{Condition the search in a variety of ways. For example, you
can limit the results to a specific number of items, control case
sensitivity when matching string attributes, etc. See 'Search parameters'
below.}

\item{update}{Named list specifying the new values of attributes.}

\item{password}{Password to unlock the keychain, or new password to
set when creating a new keychain. May be \code{NULL} in interactive
sessions, to force a secure password dialog.}
}
\description{
\verb{macos_item_*} functions add, delete, update and search Keychain items.

\verb{macos_keychain_*} functions create, delete, list, lock, unlock
keychains.
}
\details{
\subsection{Keychain items}{

TODO
\subsection{Attributes for different Keychain item classes}{

TODO
}

}

\subsection{Keychains}{

TODO
}

\subsection{Search Parameters}{

TODO
}
}
